[Feat] #7 - RentView 추가
